third_party/WebKit/LayoutTests/external/wpt/safe-html-types/document-write.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./support/helper.js"></script>
<body>
<script>
  promise_test(t => {
    return SafeHTML.escape(unescapedHTML).then(html => {
      return createFrameAndWrite(html);
    }).then(i => {
      assert_equals(i.contentDocument.body.innerText, unescapedHTML, "innerText");
      i.remove();
    });
  }, "document.write(SafeHTML).");
</script>
